// src/redux/types/inventoryType.ts
//---------------Inventory List---------------
export enum listActionType {
  LIST_SUCCESS = 'LIST_SUCCESS',
  LIST_FAIL = 'LIST_FAIL'
}

//---------------Search List---------------
export enum searchListActionType {
  SEARCHLIST_SUCCESS = 'SEARCHLIST_SUCCESS',
  SEARCHLIST_FAIL = 'SEARCHLIST_FAIL'
}
//---------------Search product List---------------
export enum searchProductActionType {
  SEARCH_PRODUCT_SUCCESS = 'SEARCH_PRODUCT_SUCCESS',
  SEARCH_PRODUCT_FAIL = 'SEARCH_PRODUCT_FAIL'
}
//---------------Input Inventory--------------
export enum inputInventoryActionType {
  IMPUT_INVENTORY_SUCCESS = 'IMPUT_INVENTORY_SUCCESS',
  IMPUT_INVENTORY_FAIL = 'IMPUT_INVENTORY_FAIL'
}